[@backend-engineer] PercivalBpmEstimator sampleRate를 int로 변경 — Essentia INTEGER 타입 요구사항

+++ backend/routers/analyze.py
@@ -62,7 +62,7 @@ def _analyze(pcm_bytes: bytes, sample_rate: int) -> dict:
             frameSize=int(1024), frameSizeOSS=int(2048),
             hopSize=int(128), hopSizeOSS=int(128),
             maxBPM=float(220), minBPM=float(50),
-            sampleRate=float(sample_rate),
+            sampleRate=int(sample_rate),
         )
         bpm = bpm_est(pcm)
         result["bpm"] = float(bpm)
